providers/genai/pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
